eventloop: Introduce run_until_complete and callback
This is basically a wrapper to loop.run_until_complete or
asyncio.run_coroutine_threadsafe, depending on the context.
It allows a normal function to run a coroutine in the event loop,
provided it wasn't called from a coroutine itself.

The function callback basically calls loop.call_soon_threadsafe with the
right arguments.

Signed-off-by: SylvainGault <github@sylvaingault.fr>